Split application to python module
