--- embedded_server/smart4l_api/smart4l_api.py
@@ -1,15 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-__author__ = "cbarange"
-__license__ = "MIT"
-__date__ = "11-08-2020"
-__version__ = "1.0.0"
-__status__ = "Prototype"
-
-import os  # standard library
-import sys
-
-
-def run():
-    # DO SOMETHING
-    pass
