+++ embedded_server/smart4l_service/POC/POC_network.py
