+++ embedded_server/smart4l_service/POC/POC_websocket.html
