+++ embedded_server/smart4l_service/POC/smart4_websocket.py
