--- embedded_server/smart4l_service/globalVar.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# --- Global variable ---
-process = None  # Use to start, stop, reload app
-
-
