+++ embedded_server/smart4l_service/smart4l_socket.py
@@ -1,0 +1,29 @@
+# -*- coding: utf-8 -*-
+
+from utils import Message, Status, ServiceObjectInterface
+
+
+
+class Smart4lServeur(ServiceObjectInterface):
+    def __init__(self):
+        pass
+
+    def do(self):
+        pass
+
+    def stop(self):
+        pass
+
+
+
+
+class Smart4lClient(ServiceObjectInterface):
+    def __init__(self):
+        pass
+
+    def do(self):
+        pass
+
+    def stop(self):
+        pass
+
